moved action.js file to src folder and changed path in itemGrid
